chore: create score domain
